backend/services/admin_reporting_service/pyproject.toml
[project]
name = "admin-reporting-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.5.3",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "whitenoise>=6.9.0",
]
